+++ plugins/postcss-is-pseudo-class/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to PostCSS Is Pseudo Class
 
+### Unreleased (major)
+
+- Updated: Support for Node `20.19.0` or later (major).
+- Removed: `commonjs` API. In supported Node versions `require(esm)` will work without needing to make code changes.
+
 ### 5.0.3
 
 _June 11, 2025_