+++ plugins/postcss-is-pseudo-class/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Updated: Support for Node `20.19.0` or later (major).
 - Removed: `commonjs` API. In supported Node versions `require(esm)` will work without needing to make code changes.
+- Updated [`@csstools/selector-specificity`](https://github.com/csstools/postcss-plugins/tree/main/packages/selector-specificity) to [`6.0.0`](https://github.com/csstools/postcss-plugins/tree/main/packages/selector-specificity/CHANGELOG.md#600) (major)
 
 ### 5.0.3